Move getData and postData functions out of config and into utilities file
