+++ app/client/src/config.tsx
@@ -71,40 +71,3 @@ export const messages = {
   paymentRequestFormWillBeDeleted:
     "A request to edit the Application form associated with this draft or submitted Payment Request form has been made, so this form has been set to read-only mode. Visit your dashboard to make edits to the associated Application form submission.",
 };
-
-async function fetchData<T = any>(url: string, options: RequestInit) {
-  try {
-    const response = await fetch(url, options);
-    if (!response.ok) throw new Error(response.statusText);
-    const contentType = response.headers.get("content-type");
-    return contentType?.includes("application/json")
-      ? ((await response.json()) as Promise<T>)
-      : (Promise.resolve() as Promise<T>);
-  } catch (error) {
-    return await Promise.reject(error);
-  }
-}
-
-/**
- * Fetches data and returns a promise containing JSON fetched from a provided
- * web service URL or handles any other OK response returned from the server
- */
-export function getData<T = any>(url: string) {
-  return fetchData<T>(url, {
-    method: "GET",
-    credentials: "include" as const,
-  });
-}
-
-/**
- * Posts JSON data and returns a promise containing JSON fetched from a provided
- * web service URL or handles any other OK response returned from the server
- */
-export function postData<T = any>(url: string, data: object) {
-  return fetchData<T>(url, {
-    method: "POST",
-    credentials: "include" as const,
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(data),
-  });
-}
