+++ app/client/src/routes/paymentRequestForm.tsx
@@ -6,8 +6,10 @@ import { Formio, Form } from "@formio/react";
 import { cloneDeep, isEqual } from "lodash";
 import icons from "uswds/img/sprite.svg";
 // ---
-import { serverUrl, messages, getData, postData } from "../config";
+import { serverUrl, messages } from "../config";
 import {
+  getData,
+  postData,
   useCsbData,
   useBapSamData,
   useSubmissionsQueries,
